- more fonts - mouse movement support (no buttons yet)

+++ shapes/point.cpp
@@ -5,7 +5,7 @@ Point::Point(){
     y = 0;
 }
 
-Point::Point(long _x, long _y){
+Point::Point(int _x, int _y){
     x = _x;
     y = _y;
 };
